Move MaybeNext code into a new "offset" submodule
Tested-by: Kristofer Rye <kristofer.rye@gmail.com>
